backend/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
Django==4.2.5
django-htmlmin==0.11.0
djangorestframework==3.14.0
html5lib==1.1
psycopg2-binary==2.9.7
pytz==2023.3.post1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.8.0
webencodings==0.5.1
